chore: folders-action-sheet cleanup and rename to folder-action-sheet
